+++ src/Pages/Cart/Cart.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { Header, Loaders } from '../../Components';
+import { Checkout, Header, Loaders } from '../../Components';
 import { HorizontalCard } from '../../Components/Horizontal_Card/HorizontalCard';
 import { useAuth, useCart } from '../../Context';
 import PriceContainer from './Components/Price Container/PriceContainer';
@@ -14,6 +14,7 @@ const Cart = () => {
 	const { authState } = useAuth();
 	const [couponType, setCouponType] = useState('');
 	const [isLoading, setIsLoading] = useState(true);
+	const [checkout, setCheckout] = useState(false);
 
 	const { cartState, cartDispatch } = useCart();
 
@@ -45,46 +46,50 @@ const Cart = () => {
 		<>
 			<Header />
 			{isLoading && <Loaders />}
-			<main className={styles.cartMain}>
-				{cartState.cart.length > 0 && !isLoading ? (
-					<>
-						<div>
-							<h3 className={styles.cartHeading}>My Cart</h3>
-							<small>
-								{cartState.cart.length} {cartState.cart.length > 1 ? 'items' : 'item'}
-							</small>
-						</div>
-						<section className={styles.cartContainer}>
+			{checkout ? (
+				<Checkout couponType={couponType} setCheckout={setCheckout} deleteHandler={deleteHandler} />
+			) : (
+				<main className={styles.cartMain}>
+					{cartState.cart.length > 0 && !isLoading ? (
+						<>
 							<div>
-								{cartState.cart.map((card) => (
-									<HorizontalCard
-										{...card}
-										key={card._id}
-										setCouponType={setCouponType}
-										deleteHandler={deleteHandler}
-									/>
-								))}
+								<h3 className={styles.cartHeading}>My Cart</h3>
+								<small>
+									{cartState.cart.length} {cartState.cart.length > 1 ? 'items' : 'item'}
+								</small>
 							</div>
+							<section className={styles.cartContainer}>
+								<div>
+									{cartState.cart.map((card) => (
+										<HorizontalCard
+											{...card}
+											key={card._id}
+											setCouponType={setCouponType}
+											deleteHandler={deleteHandler}
+										/>
+									))}
+								</div>
 
-							<PriceContainer
-								couponType={couponType}
-								setCouponType={setCouponType}
-								deleteHandler={deleteHandler}
-							/>
-						</section>
-					</>
-				) : (
-					<>
-						{!isLoading && (
-							<div className={styles.emptyCartContainer}>
-								<span className={styles.emptyMsgTitle}>Looks like your cart is empty</span>
-								<img src={emptyCart} alt="responsive image" className="resp-img" />
-								<Link to="/products"> Browse Products</Link>
-							</div>
-						)}
-					</>
-				)}
-			</main>
+								<PriceContainer
+									couponType={couponType}
+									setCouponType={setCouponType}
+									setCheckout={setCheckout}
+								/>
+							</section>
+						</>
+					) : (
+						<>
+							{!isLoading && (
+								<div className={styles.emptyCartContainer}>
+									<span className={styles.emptyMsgTitle}>Looks like your cart is empty</span>
+									<img src={emptyCart} alt="responsive image" className="resp-img" />
+									<Link to="/products"> Browse Products</Link>
+								</div>
+							)}
+						</>
+					)}
+				</main>
+			)}
 		</>
 	);
 };
